Relocate container_monitor from galaxy_ext, but keep it runnable from galaxy_ext. Add entrypoitn galaxy-container-monitor to job_execution package.
