+++ packages/job_execution/setup.cfg
@@ -44,6 +44,7 @@ python_requires = >=3.7
 [options.entry_points]
 console_scripts =
         galaxy-set-metadata = galaxy.metadata.set_metadata:set_metadata
+        galaxy-container-monitor = galaxy.job_execution.container_monitor:main
 
 [options.packages.find]
 exclude =
